requirements.txt
python-telegram-bot==20.3
transformers==4.41.1
torch
pdfminer.six
python-pptx
nltk
scipy
python-dotenv